+++ spec/es2015_spec.rb
@@ -440,6 +440,26 @@ describe "ES2015 support" do
         must_equal 'class C {get m2() {return this.m1}; get m1() {}}'
     end
 
+    it "should auto bind methods referenced as properties" do
+      to_js('class C; def m1(x); end; def m2; m1; end; end').
+        must_equal 'class C {m1(x) {}; get m2() {return this.m1.bind(this)}}'
+    end
+
+    it "should leave alone methods that are already bound" do
+      to_js('class C; def m1(x); end; def m2; m1.bind(this); end; end').
+        must_equal 'class C {m1(x) {}; get m2() {return this.m1.bind(this)}}'
+    end
+
+    it "should leave alone methods that specify a target" do
+      to_js('class C; def m1(x); end; def m2; self.m1; end; end').
+        must_equal 'class C {m1(x) {}; get m2() {return this.m1}}'
+    end
+
+    it "should not bind methods called as methods" do
+      to_js('class C; def m1(x); end; def m2; m1(); end; end').
+        must_equal 'class C {m1(x) {}; get m2() {return this.m1()}}'
+    end
+
     it "should prefix class constants referenced in methods by class name" do
       to_js('class C; X = 1; def m; X; end; end').
         must_equal 'class C {get m() {return C.X}}; C.X = 1'